+++ R/parse_disturbance_observations.R
@@ -24,6 +24,8 @@
 #'   \item location_accuracy <dbl>
 #'   \item turtle_date <date>
 #'   \item season <int>
+#'   \item season_week <int> The number of completed weeks since fiscal year start
+#'   \item iso_week <int> The number of completed weeks since calendar year start
 #'   \item comments <chr>
 #'   \item absolute_admin_url <chr>
 #'   \item photos <list>
@@ -62,7 +64,8 @@ parse_disturbance_observations <- function(wastd_api_response) {
       datetime = purrr::map_chr(., c("properties", "encounter", "properties", "when")) %>% httpdate_as_gmt08(),
       turtle_date = datetime %>% datetime_as_turtle_date(),
       season = datetime %>% datetime_as_season(),
-      week = datetime %>% datetime_as_isoweek(),
+      season_week = datetime %>% datetime_as_seasonweek(),
+      iso_week = datetime %>% datetime_as_isoweek(),
       longitude = purrr::map_dbl(., c("properties", "encounter", "properties", "longitude")),
       latitude = purrr::map_dbl(., c("properties", "encounter", "properties", "latitude")),
       crs = purrr::map_chr(., c("properties", "encounter", "properties", "crs")),