+++ R/parse_turtle_nest_encounters.R
@@ -19,7 +19,8 @@
 #'   \item datetime <dttm>
 #'   \item turtle_date <date>
 #'   \item season <int>
-#'   \item week <int>
+#'   \item season_week <int> The number of completed weeks since fiscal year start
+#'   \item iso_week <int> The number of completed weeks since calendar year start
 #'   \item longitude <dbl>
 #'   \item latitude <dbl>
 #'   \item crs <chr>
@@ -71,8 +72,8 @@ parse_turtle_nest_encounters <- function(wastd_api_response) {
       datetime = purrr::map_chr(., c("properties", "when")) %>% httpdate_as_gmt08(),
       turtle_date = datetime %>% datetime_as_turtle_date(),
       season = datetime %>% datetime_as_season(),
-      week = datetime %>% datetime_as_isoweek(),
-
+      season_week = datetime %>% datetime_as_seasonweek(),
+      iso_week = datetime %>% datetime_as_isoweek(),
 
       longitude = purrr::map_dbl(., c("properties", "longitude")),
       latitude = purrr::map_dbl(., c("properties", "latitude")),