+++ tests/testthat/test_datetime_as_isoweek.R
@@ -5,7 +5,7 @@ test_that("datetime_as_isoweek returns numeric", {
   testthat::expect_equal(class(dd), "numeric")
 })
 
-test_that("datetime_as_isoweek returns correct year at edge cases", {
+test_that("datetime_as_isoweek returns correct week at edge cases", {
   testthat::expect_equal(datetime_as_isoweek(httpdate_as_gmt08("2017-06-30T15:59:59Z")), 26)
   testthat::expect_equal(datetime_as_isoweek(httpdate_as_gmt08("2017-06-30T16:00:00Z")), 26)
   testthat::expect_equal(datetime_as_isoweek(httpdate_as_gmt08("2017-08-30T06:38:43Z")), 35)